refactoring: eigenes util-modul für EnumerateChecked
