+++ src/util.rs
@@ -2,6 +2,7 @@
 
 #[path = "util/eingeschränkt.rs"]
 pub mod eingeschränkt;
+pub mod enumerate_checked;
 pub mod nachschlagen;
 pub mod unicase_ord;
 pub mod void;
